+++ homework/src/HW1/submissions/partA/vincentsaw/uber/NotificationManager.java
@@ -1,0 +1,8 @@
+package HW1.submissions.partA.vincentsaw.uber;
+
+public class NotificationManager {
+    void sendNotification(User user, String message) {
+        // Code for sending notifications to the user
+        System.out.println("Notification sent to user: " + user.getUsername() + " - " + message);
+    }
+}
