+++ homework/src/HW1/submissions/partA/vincentsaw/uber/README.md
@@ -1,5 +1,7 @@
 ### (1) Violated SOLID Principles
-Lorem ipsum
+This company violates the **Single Responsibility Principle**. This is because RideManager handles both fare calculations and also notifications.
 
 ### (2) Violation Fixes
-Lorem ipsum
+To fix this principle's violation, we can split the responsibility of RideManager up.
+
+Not strictly related to the Single Responsibility Principle, but we can also split the Ride and User classes into different files to allow for greater modularity in terms of access levels.
