+++ homework/src/HW1/submissions/partA/vincentsaw/uber/Ride.java
@@ -1,0 +1,28 @@
+package HW1.submissions.partA.vincentsaw.uber;
+
+public class Ride {
+    private double distanceInMiles;
+    private int durationInMinutes;
+
+    public Ride(double distanceInMiles, int durationInMinutes) {
+        this.distanceInMiles = distanceInMiles;
+        this.durationInMinutes = durationInMinutes;
+    }
+
+    public double getDistanceInMiles() {
+        return distanceInMiles;
+    }
+
+    public void setDistanceInMiles(double distanceInMiles) {
+        this.distanceInMiles = distanceInMiles;
+    }
+
+    public int getDurationInMinutes() {
+        return durationInMinutes;
+    }
+
+    public void setDurationInMinutes(int durationInMinutes) {
+        this.durationInMinutes = durationInMinutes;
+    }
+
+}
