+++ homework/src/HW1/submissions/partA/vincentsaw/uber/User.java
@@ -1,0 +1,16 @@
+package HW1.submissions.partA.vincentsaw.uber;
+
+public class User {
+    // User details
+    private String username;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+
+}
